requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
python-dotenv
deepface
tensorflow
opencv-python-headless
Pillow
numpy